chore(package): package functions individually - EUBFR-35
